set params::validate default value of plugin_dir to undef.
Without this patch the bacula class's default value of undef, will not be used, since undef does not overwrite a parameter.
Thus the params::validate default value for plugin_dir will be used, which is ''.
This causes the validation to fail since '' is not a valid and absolute path.
To fix this behaviour, the default value for plugin_dir within the params::validate class should also be undef, to work
with the != undef condition within the class.
The error will be triggered if a bacula class is created without setting the plugin_dir argument.

The resulting error message is:
valuation Error: Error while evaluating a Function Call,  is not an absolute path.